refactored home page to use i18n, added home in types
