+++ app/i18n/en/home.ts
@@ -1,0 +1,29 @@
+import { type HomeType } from '../types/home';
+
+const home: HomeType = {
+  introSection: [
+    'Welcome to Virufy',
+    'Your Digital Health Companion',
+    'Empower yourself with the latest in health technology.',
+    'Introducing Virufy, the cutting-edge app that utilizes advanced audio and generative AI algorithms to analyze your breathing patterns and provide insights into potential infection symptoms.',
+    'Demo App',
+    '*Any device with a web browser and microphone (e.g. iPhone, Android, laptop, iPad).',
+    '*The app is still under development, support us by contributing your data into our data collection app.',
+    'How It Works',
+    "With just a simple intentional cough into your device's microphone, Virufy harnesses the power of artificial intelligence to detect unique sound patterns associated with respiratory diseases, such as COVID-19, TB, flu, RSV, COPD, and asthma.",
+  ],
+  section2: [
+    'Your Health, Our Priority',
+    'At Virufy, your well-being is at the heart of everything we do. We prioritize privacy, ensuring that your data remains secure and confidential.',
+    'Instant Analysis',
+    "Receive immediate prompt feedback on your cough's characteristics and abnormality profile based on the sound pattern of your cough*",
+    'At-home Anonymity',
+    'Feel secure knowing that your results are only for your eyes.',
+    'User-Friendly Interface',
+    'Intuitive design for easy navigation and seamless user experience.',
+    '*The results of the analysis done by the system will not replace the diagnosis by a specialized doctor.',
+    'Our Technology',
+  ],
+};
+
+export default home;
