+++ app/i18n/es/home.ts
@@ -1,0 +1,29 @@
+import { type HomeType } from '../types/home';
+
+const home: HomeType = {
+  introSection: [
+    'Bienvenido a Virufy',
+    'Tu Compañero de Salud Digital',
+    'Empodérate con lo último en tecnología de salud.',
+    'Presentamos Virufy, la aplicación de vanguardia que utiliza algoritmos avanzados de audio e IA generativa para analizar tus patrones de respiración y proporcionar información sobre posibles síntomas de infección.',
+    'Demo de la App',
+    '*Cualquier dispositivo con un navegador web y micrófono (por ejemplo, iPhone, Android, portátil, iPad).',
+    '*La aplicación aún está en desarrollo, apóyanos contribuyendo con tus datos en nuestra aplicación de recolección de datos.',
+    'Cómo Funciona',
+    "Con solo una tos intencional hacia el micrófono de tu dispositivo, Virufy aprovecha el poder de la inteligencia artificial para detectar patrones de sonido únicos asociados con enfermedades respiratorias, como COVID-19, tuberculosis, gripe, VSR, EPOC y asma.",
+  ],
+  section2: [
+    'Tu Salud, Nuestra Prioridad',
+    'En Virufy, tu bienestar es el centro de todo lo que hacemos. Priorizamos la privacidad, asegurando que tus datos permanezcan seguros y confidenciales.',
+    'Análisis Instantáneo',
+    'Recibe retroalimentación inmediata sobre las características de tu tos y el perfil de anormalidad basado en el patrón de sonido de tu tos*',
+    'Anonimato en Casa',
+    'Siéntete seguro sabiendo que tus resultados son solo para tus ojos.',
+    'Interfaz Fácil de Usar',
+    'Diseño intuitivo para una fácil navegación y una experiencia de usuario fluida.',
+    '*Los resultados del análisis realizado por el sistema no reemplazarán el diagnóstico de un médico especializado.',
+    'Nuestra Tecnología',
+  ],
+};
+
+export default home;
