+++ app/i18n/es/index.ts
@@ -3,6 +3,7 @@ import ai from './ai';
 import amilsStory from './amilsStory';
 import faq from './faq';
 import footer from './footer';
+import home from './home';
 import jobDetails from './jobDetails';
 import jobListing from './jobListing';
 import navbar from './navbar';
@@ -19,6 +20,7 @@ const esMetadata: I18nData = {
   amilsStory,
   faq,
   footer,
+  home,
   jobDetails,
   jobListing,
   navbar,
