+++ app/i18n/ja/home.ts
@@ -1,0 +1,29 @@
+import { type HomeType } from '../types/home';
+
+const home: HomeType = {
+  introSection: [
+    'Virufyへようこそ',
+    'あなたのデジタルヘルスコンパニオン',
+    '最新のヘルステクノロジーで自分を強化しましょう。',
+    'Virufyは、あなたの呼吸パターンを分析し、感染症の症状に関する洞察を提供するために、先進的な音声解析と生成的AIアルゴリズムを活用した最先端のアプリです。',
+    'デモアプリ',
+    '*Webブラウザとマイクを備えた任意のデバイス（例：iPhone、Android、ノートパソコン、iPad）。',
+    '*アプリはまだ開発中です。データ収集アプリにあなたのデータを提供して私たちをサポートしてください。',
+    '使い方',
+    'デバイスのマイクに向かって意図的に咳をするだけで、Virufyは人工知能の力を利用して、COVID-19、結核、インフルエンザ、RSウイルス、COPD、喘息などの呼吸器疾患に関連する独自の音声パターンを検出します。',
+  ],
+  section2: [
+    'あなたの健康が私たちの優先事項',
+    'Virufyでは、あなたの健康がすべての中心です。私たちはプライバシーを最優先し、あなたのデータが安全で機密性を保つことを保証します。',
+    '即時分析',
+    '咳の音声パターンに基づいて、咳の特徴と異常プロファイルに関する即時のフィードバックを受け取ります*',
+    '自宅での匿名性',
+    '結果はあなただけが確認できるため、安心してください。',
+    'ユーザーフレンドリーなインターフェース',
+    '直感的なデザインで、簡単に操作できるスムーズなユーザー体験を提供します。',
+    '*システムによる分析結果は、専門医による診断の代わりにはなりません。',
+    '私たちの技術',
+  ],
+};
+
+export default home;
