+++ app/i18n/types/home.ts
@@ -1,0 +1,4 @@
+export interface HomeType {
+  introSection: Array<string>;
+  section2: Array<string>;
+}
